Rolled in Cash's edits and some further bug fixes
git-svn-id: http://code.elgg.org/plugins/plugins@5052 36083f99-b078-4883-b0ff-0f9b5a30f544

+++ messages/actions/delete.php
@@ -29,7 +29,7 @@
 	        $message = get_entity($message_id);
 	        
 	    // Make sure we actually have permission to edit and that the object is of sub-type messages
-			if ($message->getSubtype() == "messages") {
+			if ($message && $message->getSubtype() == "messages") {
 	    		
 				if ($submit == elgg_echo('delete')) {
 					if ($message->delete()) {
@@ -65,8 +65,8 @@
     		    forward("mod/messages/?username=" . $_SESSION['user']->username . "&offset={$offset}");
 		    }
         } else {
-        	register_error(elgg_echo("messages:nopermission"));
-        	forward("mod/messages/?offset={$offset}&username=" . $_SESSION['user']->username);
+        	register_error(elgg_echo("messages:notfound"));
+        	forward($_SERVER['HTTP_REFERER']);
         }
